requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==6.0.1
Werkzeug==3.1.3
python-dotenv==1.1.1
groq==0.9.0
gunicorn==21.2.0